Updated to support linux-g++-64 mkspec

+++ src/libwad.cpp
@@ -49,7 +49,7 @@ QStringList DoomWad::getMapNames() {
 
 		if (wad.read((char*)&header, sizeof(wadinfo_t))==sizeof(wadinfo_t)&&wad.seek(header.infotableofs)) {
 			filelump_t *fileinfo=new filelump_t[header.numlumps];
-			size_t length=sizeof(filelump_t)*header.numlumps;
+            qint64 length=sizeof(filelump_t)*header.numlumps;
 
 			if (wad.read((char*)fileinfo, length)==length) {
 				char prev_lump_name[9]={};
@@ -85,7 +85,7 @@ QString DoomWad::getIwadinfoName()
 
 		if (wad.read((char*)&header, sizeof(wadinfo_t))==sizeof(wadinfo_t)&&wad.seek(header.infotableofs)) {
 			filelump_t *fileinfo=new filelump_t[header.numlumps];
-			size_t length=sizeof(filelump_t)*header.numlumps;
+            qint64 length=sizeof(filelump_t)*header.numlumps;
 
 			if (wad.read((char*)fileinfo, length)==length) {
 				for (int i=0; i<header.numlumps; i++) {
@@ -124,7 +124,7 @@ bool DoomWad::isMAPXX()
 
 		if (wad.read((char*)&header, sizeof(wadinfo_t))==sizeof(wadinfo_t)&&wad.seek(header.infotableofs)) {
 			filelump_t *fileinfo=new filelump_t[header.numlumps];
-			size_t length=sizeof(filelump_t)*header.numlumps;
+            qint64 length=sizeof(filelump_t)*header.numlumps;
 
 			if (wad.read((char*)fileinfo, length)==length) {
 				for (int i=0; i<header.numlumps; i++) {
